// src/app/modules/shared/classes/helper/enum.ts
export enum ContextType {
    TASK = 1,
    OBJECTIVE = 2,
}
export enum ContartType {
    YEARLY = 1,
    SECONDARY = 2,
}

export enum ROLE_LEVELS {
    SuperAdmin = 0,
    MasterAdmin = 1,
    CountryAdmin = 2,
    AreaManager = 3,
    AreaInCharge = 4,
    SalesMan = 5,
    Merchandiser = 6,
    CashVan = 7,
    MasterUploader = 8,
    CountryUploader = 9,
    TradeMarketer = 10,
    Promoter = 11,
}


export enum ROLE_IDS {
    SuperAdmin = "1",
    AssetOwnerAdmin = "2",
    ContractorAdmin = "3",
    User = "4",
    TaskSubmitter = "5",
    Viewer = "6",
}